+++ include/lambdamart/treelearner.h
@@ -122,7 +122,7 @@ class TreeLearner {
 
 public:
     TreeLearner() = delete;
-    TreeLearner(const Dataset* _dataset, const double* _gradients, const double* _hessians, const Config* _config) :
+    TreeLearner(const Dataset* _dataset, const float* _gradients, const float* _hessians, const Config* _config) :
         dataset(_dataset), gradients(_gradients), hessians(_hessians), config(_config)
     {
         tie(num_samples, num_features) = dataset->shape();
@@ -162,8 +162,8 @@ private:
     // as input
     const Config*             config;
     const Dataset*            dataset;
-    const double*             gradients;
-    const double*             hessians;
+    const float*             gradients;
+    const float*             hessians;
 
     // as working set
     sample_t                            num_samples;
@@ -173,12 +173,12 @@ private:
     std::vector<SplitInfo>              best_splits;
     size_t                              max_splits;
     sample_t                            min_data_in_leaf;
-    std::vector<double>                 node_to_output;
-    std::vector<unsigned int>           sample_to_node;
+    std::vector<float>                 node_to_output;
+    std::vector<uint16_t>           sample_to_node;
     std::vector<SplitCandidate*>        split_candidates;
     std::vector<NodeStats*>             node_info;
-    std::vector<int>                    node_to_candidate;
-    std::vector<int>                    sample_to_candidate;  // -1: this sample doesn't exist in any candidate node
+    std::vector<uint16_t>                    node_to_candidate;
+    std::vector<int16_t>                    sample_to_candidate;  // -1: this sample doesn't exist in any candidate node
     nodeidx_t                           num_candidates = 0;
     std::priority_queue<SplitCandidate*, std::vector<SplitCandidate*>, CmpCandidates> node_queue;
 
@@ -187,7 +187,7 @@ private:
     bool   select_split_candidates();
     void   find_best_splits();
     void   perform_split();
-    double get_sample_score(sample_t sid) { return node_to_output[sample_to_node[sid]]; };
+    float get_sample_score(sample_t sid) { return node_to_output[sample_to_node[sid]]; };
 };
 
 }
